+++ examples/03-tiles/index.html
@@ -180,7 +180,7 @@ async function main () {
     })
 
      const tileNodeBG = await Cobalt.initNode(Game.renderer, {
-        type: 'cobalt:tile',
+        type: 'cobalt:tileHDR',
         refs: {
             // key is the var name defined in this node
             // value is the var name in the cobalt resources dictionary
@@ -196,7 +196,7 @@ async function main () {
 
     // instantiate all runnable nodes
     const tileNodeBg = await Cobalt.initNode(Game.renderer, {
-        type: 'cobalt:tile',
+        type: 'cobalt:tileHDR',
         refs: {
             // key is the var name defined in this node
             // value is the var name in the cobalt resources dictionary
@@ -210,7 +210,7 @@ async function main () {
     })
 
     const tileNodeFg = await Cobalt.initNode(Game.renderer, {
-        type: 'cobalt:tile',
+        type: 'cobalt:tileHDR',
         refs: {
             // key is the var name defined in this node
             // value is the var name in the cobalt resources dictionary
