--- examples/04-overlay/deps.js
@@ -1,1 +1,0 @@
-export { mat4, vec2, vec3, vec4 } from 'https://wgpu-matrix.org/dist/3.x/wgpu-matrix.module.js'
