+++ examples/05-bloom/index.html
@@ -131,7 +131,7 @@ async function main () {
 
 
     const spriteNode = await Cobalt.initNode(Game.renderer, {
-        type: 'cobalt:sprite2',
+        type: 'cobalt:spriteHDR',
         refs: {
             // key is the var name defined in this node
             // value is the var name in the cobalt resources dictionary
