+++ src/sprite-hdr/sprite.wgsl
