--- src/sprite/constants.js
@@ -1,1 +1,0 @@
-export const FLOAT32S_PER_SPRITE = 12 // vec2(translate) + vec2(scale) + vec4(tint) + opacity + rotation + emissiveIntensity + sortValue
