+++ src/spritesheet/create-sprite-quads.js
